# docs/specs/swagger.yaml
{"openapi":"3.0.1","info":{"title":"fernandez-federico","description":"project.description","contact":{"name":"fernandez-federico","email":"federico.fernandezlescano@mercadolibre.com"},"version":"project.version"},"servers":[{"url":"https://null_fernandez-federico.furyapps.io","description":"Scope null"}],"paths":{"/api/v1/users/signUp":{"post":{"tags":["user-controller"],"operationId":"signUp","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationUserDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK"}}}},"/api/v1/parts":{"post":{"tags":["part-controller"],"operationId":"addPartCountryDealerStock","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CountryDealerStockDTO"}}}},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CountryDealerStockResponseDTO"}}}}}}},"/api/v1/parts/orders":{"get":{"tags":["part-controller"],"operationId":"getSubsidiaryOrdersByDeliveryStatus","parameters":[{"name":"dealerNumber","in":"query","required":true,"schema":{"maxLength":4,"minLength":4,"type":"string"}},{"name":"deliveryStatus","in":"query","required":false,"schema":{"pattern":"\\b[PCDF]\\b","type":"string"}},{"name":"order","in":"query","required":false,"schema":{"pattern":"\\b[12]\\b","type":"string"}},{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SubsidiaryOrdersByDeliveryStatusDTO"}}}}}},"post":{"tags":["part-controller"],"operationId":"addBillCountryDealer","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BillRequestDTO"}}}},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BillDTO"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"operationId":"ping","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/parts/list":{"get":{"tags":["part-controller"],"operationId":"getList","parameters":[{"name":"params","in":"query","required":true,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PartDTO"}}}}}}}},"/api/v1/orders/list":{"get":{"tags":["bill-controller"],"operationId":"getOrderDetails","parameters":[{"name":"orderNumberCM","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BillDTO"}}}}}}}},"components":{"schemas":{"ApiError":{"type":"object","properties":{"error":{"type":"string"},"message":{"type":"string"},"status":{"type":"integer","format":"int32"}}},"ApplicationUserDTO":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"country":{"type":"string"},"role":{"type":"string"}}},"CountryDealerStockDTO":{"type":"object","properties":{"partCode":{"maxLength":8,"minLength":8,"type":"string"},"quantity":{"minimum":1,"type":"integer","format":"int32"}}},"CountryDealerStockResponseDTO":{"type":"object","properties":{"part":{"$ref":"#/components/schemas/PartDTO"}}},"PartDTO":{"type":"object","properties":{"partCode":{"type":"string"},"description":{"type":"string"},"maker":{"type":"string"},"quantity":{"type":"integer","format":"int32"},"discountType":{"type":"string"},"normalPrice":{"type":"number","format":"double"},"urgentPrice":{"type":"number","format":"double"},"netWeight":{"type":"integer","format":"int32"},"longDimension":{"type":"integer","format":"int32"},"widthDimension":{"type":"integer","format":"int32"},"tallDimension":{"type":"integer","format":"int32"},"lastModification":{"type":"string"}}},"BillRequestDTO":{"required":["billDetails"],"type":"object","properties":{"subsidiaryNumber":{"maxLength":4,"minLength":4,"type":"string"},"deliveryDate":{"type":"string","format":"date"},"billDetails":{"type":"array","items":{"$ref":"#/components/schemas/PostBillDetailDTO"}}}},"PostBillDetailDTO":{"type":"object","properties":{"partCode":{"maxLength":8,"minLength":8,"type":"string"},"quantity":{"minimum":1,"type":"integer","format":"int32"},"accountType":{"type":"string","enum":["G","R"]}}},"BillDTO":{"type":"object","properties":{"orderNumber":{"type":"string"},"orderDate":{"type":"string","format":"date"},"deliveryDate":{"type":"string","format":"date"},"daysDelayed":{"type":"integer","format":"int32"},"deliveryStatus":{"type":"string","enum":["Procesando","Demorado","Finalizado","Cancelado"]},"orderDetails":{"type":"array","items":{"$ref":"#/components/schemas/BillDetailDTO"}}}},"BillDetailDTO":{"type":"object","properties":{"partCode":{"type":"string"},"description":{"type":"string"},"quantity":{"type":"integer","format":"int32"},"accountType":{"type":"string","enum":["G","R"]},"reason":{"type":"string"},"partStatus":{"type":"string","enum":["N","D"]}}},"SubsidiaryOrdersByDeliveryStatusDTO":{"type":"object","properties":{"subsidiaryNumber":{"type":"string"},"orders":{"type":"array","items":{"$ref":"#/components/schemas/BillDTO"}}}}}}}